Replaced vad with firevad package.
