+++ package-firefox-extension.sh
@@ -15,7 +15,7 @@ cp public/*.wasm $dest_public
 cp public/saypi.user.js $dest_public
 cp public/silero_vad.onnx $dest_public
 cp public/ort-wasm-simd-threaded.mjs $dest_public
-cp public/vad.worklet.bundle.min_firefox.js $dest_public/vad.worklet.bundle.min.js
+cp public/vad.worklet.bundle.min.js $dest_public/vad.worklet.bundle.min.js
 cp public/audio/*.mp3 $dest_public
 mkdir -p $dest_icons/logos
 cp public/icons/*.svg $dest_icons
