+++ webpack.config.js
@@ -103,7 +103,7 @@ export default (env, argv) => {
       new CopyPlugin({
         patterns: [
           {
-            from: "node_modules/@ricky0123/vad-web/dist/vad.worklet.bundle.min.js",
+            from: "node_modules/@jcrawley_dev/firevad/dist/vad.worklet.bundle.min.js",
             to: "[name][ext]",
           },
           {
